BurstTriggerI8 fix & various data type updates

+++ source/TS.NET.Tests/Triggers/FallingEdgeTriggerI8Tests.cs
@@ -45,13 +45,13 @@ namespace TS.NET.Tests
             FallingEdgeTriggerI8 trigger = new(situation.Parameters);
             trigger.SetHorizontal(situation.WindowWidth, situation.WindowTriggerPosition, situation.AdditionalHoldoff);
 
-            Span<uint> captureEndIndices = new uint[10000];
+            Span<int> captureEndIndices = new int[10000];
             var currentWindowEndIndices = captureEndIndices;
 
             for (int i = 0; i < situation.ChunkCount; i++)
             {
-                trigger.Process(situation.Input.Span.Slice((int)(i * situation.ChunkSize), (int)situation.ChunkSize), currentWindowEndIndices, out uint windowEndCount);
-                currentWindowEndIndices = currentWindowEndIndices.Slice((int)windowEndCount);
+                trigger.Process(situation.Input.Span.Slice((i * situation.ChunkSize), situation.ChunkSize), currentWindowEndIndices, out int windowEndCount);
+                currentWindowEndIndices = currentWindowEndIndices.Slice(windowEndCount);
             }
 
             for (int i = 0; i < situation.ExpectedWindowEndIndices.Length; i++)
