+++ source/TS.NET.Tests/Triggers/Situations/RisingEdgeTriggerSituations.cs
@@ -5,15 +5,15 @@ namespace TS.NET.Tests
     internal class EdgeTriggerSituation
     {
         public EdgeTriggerParameters Parameters;
-        public uint WindowWidth;
-        public uint WindowTriggerPosition;
-        public uint AdditionalHoldoff;
+        public long WindowWidth;
+        public long WindowTriggerPosition;
+        public long AdditionalHoldoff;
 
         public int ChunkSize;
         public int ChunkCount;
 
         public Memory<sbyte> Input;
-        public Memory<uint> ExpectedWindowEndIndices;
+        public Memory<int> ExpectedWindowEndIndices;
     }
 
     internal class RisingEdgeTriggerSituations
@@ -34,7 +34,7 @@ namespace TS.NET.Tests
                 ChunkCount = 1
             };
             situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-            situation.ExpectedWindowEndIndices = new uint[1];
+            situation.ExpectedWindowEndIndices = new int[1];
             situation.Input.Span.Fill(0);
             situation.Input.Span.Slice(100, 100).Fill(sbyte.MaxValue);
             situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -57,7 +57,7 @@ namespace TS.NET.Tests
                 ChunkCount = 1
             };
             situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-            situation.ExpectedWindowEndIndices = new uint[1];
+            situation.ExpectedWindowEndIndices = new int[1];
             situation.Input.Span.Fill(0);
             situation.Input.Span.Slice(100, 100).Fill(sbyte.MaxValue);
             situation.Input.Span.Slice(300, 100).Fill(sbyte.MaxValue);  // This pulse should be ignored
@@ -81,7 +81,7 @@ namespace TS.NET.Tests
                 ChunkCount = 1
             };
             situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-            situation.ExpectedWindowEndIndices = new uint[2];
+            situation.ExpectedWindowEndIndices = new int[2];
             situation.Input.Span.Fill(0);
             situation.Input.Span.Slice(100, 100).Fill(sbyte.MaxValue);
             situation.Input.Span.Slice(10200, 100).Fill(sbyte.MaxValue);
@@ -106,7 +106,7 @@ namespace TS.NET.Tests
                 ChunkCount = 1
             };
             situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-            situation.ExpectedWindowEndIndices = new uint[1];
+            situation.ExpectedWindowEndIndices = new int[1];
             situation.Input.Span.Fill(0);
             situation.Input.Span.Slice(0, 100).Fill(sbyte.MaxValue);
             situation.ExpectedWindowEndIndices.Span[0] = 0;
@@ -129,7 +129,7 @@ namespace TS.NET.Tests
                 ChunkCount = 1
             };
             situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-            situation.ExpectedWindowEndIndices = new uint[1];
+            situation.ExpectedWindowEndIndices = new int[1];
             situation.Input.Span.Fill(0);
             situation.Input.Span[100] = sbyte.MaxValue;
             situation.ExpectedWindowEndIndices.Span[0] = 10100;
