+++ source/TS.NET/Processing/Triggers/AnyEdgeTriggerI8.cs
@@ -12,11 +12,11 @@ public class AnyEdgeTriggerI8 : ITriggerI8
     private sbyte upperArmLevel;
     private sbyte lowerArmLevel;
 
-    private ulong captureSamples;
-    private ulong captureRemaining;
+    private long captureSamples;
+    private long captureRemaining;
 
-    private ulong holdoffSamples;
-    private ulong holdoffRemaining;
+    private long holdoffSamples;
+    private long holdoffRemaining;
 
     public AnyEdgeTriggerI8(EdgeTriggerParameters parameters)
     {
@@ -39,7 +39,7 @@ public class AnyEdgeTriggerI8 : ITriggerI8
         lowerArmLevel += (sbyte)parameters.Hysteresis;
     }
 
-    public void SetHorizontal(ulong windowWidth, ulong windowTriggerPosition, ulong additionalHoldoff)
+    public void SetHorizontal(long windowWidth, long windowTriggerPosition, long additionalHoldoff)
     {
         if (windowWidth < 1000)
             throw new ArgumentException($"windowWidth cannot be less than 1000");
@@ -56,12 +56,12 @@ public class AnyEdgeTriggerI8 : ITriggerI8
     }
 
     //[MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public void Process(ReadOnlySpan<sbyte> input, Span<uint> windowEndIndices, out uint windowEndCount)
+    public void Process(ReadOnlySpan<sbyte> input, Span<int> windowEndIndices, out int windowEndCount)
     {
-        uint inputLength = (uint)input.Length;
-        uint simdLength = inputLength - 32;
+        int inputLength = input.Length;
+        int simdLength = inputLength - 32;
         windowEndCount = 0;
-        uint i = 0;
+        int i = 0;
 
         Vector256<sbyte> triggerLevelVector = Vector256.Create(triggerLevel);
         Vector256<sbyte> upperArmLevelVector = Vector256.Create(upperArmLevel);
@@ -96,12 +96,12 @@ public class AnyEdgeTriggerI8 : ITriggerI8
                             }
                             while (i < inputLength)
                             {
-                                if (samplesPtr[(int)i] <= lowerArmLevel)
+                                if (samplesPtr[i] <= lowerArmLevel)
                                 {
                                     triggerState = TriggerState.ArmedRisingEdge;
                                     break;
                                 }
-                                if (samplesPtr[(int)i] >= upperArmLevel)
+                                if (samplesPtr[i] >= upperArmLevel)
                                 {
                                     triggerState = TriggerState.ArmedFallingEdge;
                                     break;
@@ -124,7 +124,7 @@ public class AnyEdgeTriggerI8 : ITriggerI8
                             }
                             while (i < inputLength)
                             {
-                                if (samplesPtr[(int)i] > triggerLevel)
+                                if (samplesPtr[i] > triggerLevel)
                                 {
                                     triggerState = TriggerState.InCapture;
                                     captureRemaining = captureSamples;
@@ -148,7 +148,7 @@ public class AnyEdgeTriggerI8 : ITriggerI8
                             }
                             while (i < inputLength)
                             {
-                                if (samplesPtr[(int)i] < triggerLevel)
+                                if (samplesPtr[i] < triggerLevel)
                                 {
                                     triggerState = TriggerState.InCapture;
                                     captureRemaining = captureSamples;
@@ -159,10 +159,10 @@ public class AnyEdgeTriggerI8 : ITriggerI8
                             break;
                         case TriggerState.InCapture:
                             {
-                                uint remainingSamples = inputLength - i;
+                                int remainingSamples = inputLength - i;
                                 if (remainingSamples > captureRemaining)
                                 {
-                                    i += (uint)captureRemaining;    // Cast is ok because remainingSamples (in the conditional expression) is uint
+                                    i += (int)captureRemaining;    // Cast is ok because remainingSamples (in the conditional expression) is uint
                                     captureRemaining = 0;
                                 }
                                 else
@@ -172,7 +172,7 @@ public class AnyEdgeTriggerI8 : ITriggerI8
                                 }
                                 if (captureRemaining == 0)
                                 {
-                                    windowEndIndices[(int)windowEndCount++] = i;
+                                    windowEndIndices[windowEndCount++] = i;
                                     if (holdoffSamples > 0)
                                     {
                                         triggerState = TriggerState.InHoldoff;
@@ -188,10 +188,10 @@ public class AnyEdgeTriggerI8 : ITriggerI8
                             break;
                         case TriggerState.InHoldoff:
                             {
-                                uint remainingSamples = inputLength - i;
+                                int remainingSamples = inputLength - i;
                                 if (remainingSamples > holdoffRemaining)
                                 {
-                                    i += (uint)holdoffRemaining;    // Cast is ok because remainingSamples (in the conditional expression) is uint
+                                    i += (int)holdoffRemaining;    // Cast is ok because remainingSamples (in the conditional expression) is uint
                                     holdoffRemaining = 0;
                                 }
                                 else
